Day 3 - Added Tree Questions(leetcode)

+++ Day-2/chef_and_max_star_value.cpp
@@ -10,15 +10,15 @@ void solve(vector<ll> &A)
     vector<ll> res(n, 0);
     unordered_set<ll> s; 
     int max_ele = INT_MIN;
-    for (int i = 0; i < n; i++) { 
+    for (ll i = 0; i < n; i++) { 
         s.insert(A[i]); 
         max_ele = max(max_ele, A[i]); 
     } 
-    for (int i = 0; i < n; i++) { 
+    for (ll i = 0; i < n; i++) { 
         if (A[i] != 0) { 
             for (int j = A[i] * 2; j <= max_ele; j += A[i]) { 
                 if (s.find(j) != s.end()) 
-                    res[i]++; 
+                    res[i]++;
             } 
         } 
     }
